Solved the wall collision bug which caused the ball to immediately relaunch due to a wall detection, as soon as the game opened. Solved this bug by modifying the wall detection code to detect IF the ball's X coordinates are *between* the walls minX and maxX bounds, for the vertical borders, and for the horizontal borders, IF the ball's Y coordinates are between the walls minY and maxY bounds. Also, a withinBorder boolean has been introduced to detect if the ball is within the border's range.
Also fixed a minor bug in the Border class where the horizontal lines where being added wrongly to the vertical borders array, and the vertical lines to the horizontal borders array. They are now added to the correct corresponding array.

A bug still exists where if the ball rebounds into a wall that is was the same side (horizontal/vertical) as the one is previously rebounded off of, it then goes through that wall.